+++ eras/mary/impl/test/Test/Cardano/Ledger/Mary/BinarySpec.hs
@@ -8,7 +8,6 @@ module Test.Cardano.Ledger.Mary.BinarySpec (spec) where
 import Cardano.Ledger.Mary
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec, txSizeSpec)
-import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra (..))
 import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Mary.Binary.Annotator ()
 import Test.Cardano.Ledger.Mary.Era ()
@@ -22,17 +21,3 @@ spec = do
   describe "DecCBOR instances equivalence" $ do
     Binary.decoderEquivalenceCoreEraTypesSpec @MaryEra
   Binary.txSizeSpec @MaryEra
-
-instance RuleListEra MaryEra where
-  type
-    EraRules MaryEra =
-      '[ "DELEG"
-       , "DELEGS"
-       , "DELPL"
-       , "LEDGER"
-       , "LEDGERS"
-       , "POOL"
-       , "PPUP"
-       , "UTXO"
-       , "UTXOW"
-       ]